demo/requirements.txt
absl-py==0.7.1
astor==0.7.1
Click==7.0
cycler==0.10.0
Flask==1.0.2
gast==0.2.2
google-pasta==0.1.5
grpcio==1.20.0
h5py==2.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.0.3
numpy==1.16.2
Pillow==6.0.0
pkg-resources==0.0.0
protobuf==3.7.1
pyparsing==2.4.0
python-dateutil==2.8.0
six==1.12.0
tb-nightly==1.14.0a20190301
tensorflow==2.0.0a0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019030115
Werkzeug==0.15.2
